requirements.txt
inferactively-pymdp==1.0.0
jax
jaxlib
numpy
matplotlib
seaborn
pandas
